v0.5 Release
Everything works for the basics. How to written. All scenes work, except
for the main caster cam. This exists but isn't finalized. Worst case
scenario its just a full screen camera for now.

As this is a beta release I have no idea if it'll be stable, or heck
even work at all.